start k8s client tests
